Proper but simple tests
* Added two definitive tests
* Should work most of the times
* Refactored out some logic to config.py
* Updated requirements
