+++ requirements.txt
@@ -5,3 +5,4 @@ pre-commit==3.6.2
 black==23.12.1
 isort==5.13.0
 autoflake==2.3.1
+pytest==8.2.0
